splitting pages from routes as the logic for this does not make sense
