+++ src/pages/LargeWork.tsx
