+++ src/pages/ResultPage.tsx
